refactoring master classes
