+++ core/server/common/src/main/java/alluxio/master/PrimarySelector.java
@@ -13,6 +13,7 @@ package alluxio.master;
 
 import alluxio.conf.Configuration;
 import alluxio.conf.PropertyKey;
+import alluxio.master.journal.ufs.UFSJournalMultiMasterPrimarySelector;
 import alluxio.util.interfaces.Scoped;
 
 import java.net.InetSocketAddress;
@@ -44,7 +45,7 @@ public interface PrimarySelector {
       String zkAddress = Configuration.getString(PropertyKey.ZOOKEEPER_ADDRESS);
       String zkElectionPath = Configuration.getString(PropertyKey.ZOOKEEPER_ELECTION_PATH);
       String zkLeaderPath = Configuration.getString(PropertyKey.ZOOKEEPER_LEADER_PATH);
-      return new PrimarySelectorClient(zkAddress, zkElectionPath, zkLeaderPath);
+      return new UFSJournalMultiMasterPrimarySelector(zkAddress, zkElectionPath, zkLeaderPath);
     }
 
     /**
@@ -55,7 +56,7 @@ public interface PrimarySelector {
       String zkElectionPath = Configuration.getString(
           PropertyKey.ZOOKEEPER_JOB_ELECTION_PATH);
       String zkLeaderPath = Configuration.getString(PropertyKey.ZOOKEEPER_JOB_LEADER_PATH);
-      return new PrimarySelectorClient(zkAddress, zkElectionPath, zkLeaderPath);
+      return new UFSJournalMultiMasterPrimarySelector(zkAddress, zkElectionPath, zkLeaderPath);
     }
 
     private Factory() {} // Not intended for instantiation.
