+++ core/server/common/src/main/java/alluxio/master/journal/ufs/UFSJournalMultiMasterPrimarySelector.java
@@ -9,19 +9,20 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.master;
+package alluxio.master.journal.ufs;
 
 import alluxio.AlluxioURI;
 import alluxio.Constants;
 import alluxio.conf.Configuration;
 import alluxio.conf.PropertyKey;
+import alluxio.master.AbstractPrimarySelector;
+import alluxio.master.ZookeeperConnectionErrorPolicy;
 
 import com.google.common.base.Preconditions;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
 import org.apache.curator.framework.recipes.leader.LeaderSelector;
 import org.apache.curator.framework.recipes.leader.LeaderSelectorListener;
-import org.apache.curator.framework.recipes.leader.Participant;
 import org.apache.curator.framework.state.ConnectionState;
 import org.apache.curator.framework.state.SessionConnectionStateErrorPolicy;
 import org.apache.curator.retry.ExponentialBackoffRetry;
@@ -30,17 +31,15 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.InetSocketAddress;
-import java.util.ArrayList;
-import java.util.List;
 import javax.annotation.concurrent.NotThreadSafe;
 
 /**
  * Masters use this client to elect a leader.
  */
 @NotThreadSafe
-public final class PrimarySelectorClient extends AbstractPrimarySelector
+public final class UFSJournalMultiMasterPrimarySelector extends AbstractPrimarySelector
     implements LeaderSelectorListener {
-  private static final Logger LOG = LoggerFactory.getLogger(PrimarySelectorClient.class);
+  private static final Logger LOG = LoggerFactory.getLogger(UFSJournalMultiMasterPrimarySelector.class);
 
   /** A constant session Id for when selector is not a leader. */
   private static final int NOT_A_LEADER = -1;
@@ -63,13 +62,13 @@ public final class PrimarySelectorClient extends AbstractPrimarySelector
   private LifecycleState mLifecycleState = LifecycleState.INIT;
 
   /**
-   * Constructs a new {@link PrimarySelectorClient}.
+   * Constructs a new {@link UFSJournalMultiMasterPrimarySelector}.
    *
    * @param zookeeperAddress the address to Zookeeper
    * @param electionPath the election path
    * @param leaderPath the path of the leader
    */
-  public PrimarySelectorClient(String zookeeperAddress, String electionPath, String leaderPath) {
+  public UFSJournalMultiMasterPrimarySelector(String zookeeperAddress, String electionPath, String leaderPath) {
     mZookeeperAddress = zookeeperAddress;
     mElectionPath = electionPath;
     if (leaderPath.endsWith(AlluxioURI.SEPARATOR)) {
@@ -103,23 +102,6 @@ public final class PrimarySelectorClient extends AbstractPrimarySelector
    */
   public String getName() {
     return mName;
-  }
-
-  /**
-   * @return the list of participants
-   */
-  public List<String> getParticipants() {
-    try {
-      List<Participant> participants = new ArrayList<>(mLeaderSelector.getParticipants());
-      List<String> results = new ArrayList<>();
-      for (Participant part : participants) {
-        results.add(part.getId());
-      }
-      return results;
-    } catch (Exception e) {
-      LOG.error(e.getMessage(), e);
-      return null;
-    }
   }
 
   /**
